moved interface settings to separate preferences tab
