+++ src/main/resources/fxml/preferences_general.fxml
@@ -20,28 +20,7 @@
 		<Insets topRightBottomLeft="12"/>
 	</padding>
 	<children>
-		<HBox spacing="6" alignment="CENTER_LEFT">
-			<Label text="%preferences.general.theme"/>
-			<ChoiceBox fx:id="themeChoiceBox" disable="${!controller.licenseHolder.validLicense}"/>
-			<Hyperlink styleClass="hyperlink-underline,hyperlink-muted" text="%preferences.general.unlockThemes" onAction="#showContributeTab" visible="${!controller.licenseHolder.validLicense}" managed="${!controller.licenseHolder.validLicense}"/>
-		</HBox>
-
-		<HBox spacing="6" alignment="CENTER_LEFT">
-			<Label text="%preferences.general.interfaceOrientation" HBox.hgrow="NEVER"/>
-			<RadioButton fx:id="nodeOrientationLtr" text="%preferences.general.interfaceOrientation.ltr" alignment="CENTER_LEFT" toggleGroup="${nodeOrientation}"/>
-			<RadioButton fx:id="nodeOrientationRtl" text="%preferences.general.interfaceOrientation.rtl" alignment="CENTER_RIGHT" toggleGroup="${nodeOrientation}"/>
-		</HBox>
-
-		<CheckBox fx:id="showMinimizeButtonCheckbox" text="%preferences.general.showMinimizeButton" visible="${controller.trayMenuInitialized}" managed="${controller.trayMenuInitialized}"/>
-
-		<CheckBox fx:id="showTrayIconCheckbox" text="%preferences.general.showTrayIcon" visible="${controller.trayMenuSupported}" managed="${controller.trayMenuSupported}"/>
-
 		<CheckBox fx:id="startHiddenCheckbox" text="%preferences.general.startHidden" />
-
-		<HBox spacing="6" alignment="CENTER_LEFT">
-			<Label text="%preferences.general.language"/>
-			<ChoiceBox fx:id="preferredLanguageChoiceBox"/>
-		</HBox>
 
 		<HBox spacing="6" alignment="CENTER_LEFT">
 			<CheckBox fx:id="debugModeCheckbox" text="%preferences.general.debugLogging"/>
