+++ src/main/resources/i18n/strings.properties
@@ -208,6 +208,8 @@ preferences.general.keychainBackend=Store passwords with
 preferences.general.interfaceOrientation=Interface Orientation
 preferences.general.interfaceOrientation.ltr=Left to Right
 preferences.general.interfaceOrientation.rtl=Right to Left
+## Interface
+preferences.interface=Interface
 ## Volume
 preferences.volume=Virtual Drive
 preferences.volume.type=Volume Type
